Refactor: Rename Animated to Avatar.
