+++ src/scenes/Activity/Ava.js
@@ -4,12 +4,12 @@ import PropTypes from 'prop-types';
 import styled from 'styled-components';
 import { Frame } from 'styles';
 
-import { Animated } from 'components';
+import { Avatar } from 'components';
 
 function Layout() {
   return (
     <Frame.Wrapper>
-      <StyledAnimated />
+      <StyledAvatar />
     </Frame.Wrapper>
   );
 }
@@ -30,7 +30,7 @@ Structure.propTypes = {
   children: PropTypes.node.isRequired
 };
 
-const StyledAnimated = styled(Animated)`
+const StyledAvatar = styled(Avatar)`
   grid-row: 26;
   align-self: end;
   justify-self: end;
